Remove old, rename new
