+++ bound/golang/crypto/jwk.go
@@ -1,12 +1,8 @@
 package crypto
 
-/*
-#cgo LDFLAGS: -L../../../target/release -lweb5_c
-#include <stdlib.h>
-#include "../../../bindings/web5_c/web5_c.h"
-*/
-import "C"
-import "unsafe"
+import (
+	"web5/web5c"
+)
 
 type JWK struct {
 	ALG string `json:"alg,omitempty"`
@@ -17,36 +13,21 @@ type JWK struct {
 	Y   string `json:"y,omitempty"`
 }
 
-func (j JWK) ToCJwk() *C.CJwk {
-	return &C.CJwk{
-		alg: C.CString(j.ALG),
-		kty: C.CString(j.KTY),
-		crv: C.CString(j.CRV),
-		d:   C.CString(j.D),
-		x:   C.CString(j.X),
-		y:   C.CString(j.Y),
-	}
-}
-
-func (j JWK) FreeCJwk(cJwk *C.CJwk) {
-	C.free(unsafe.Pointer(cJwk.alg))
-	C.free(unsafe.Pointer(cJwk.kty))
-	C.free(unsafe.Pointer(cJwk.crv))
-	C.free(unsafe.Pointer(cJwk.d))
-	C.free(unsafe.Pointer(cJwk.x))
-	C.free(unsafe.Pointer(cJwk.y))
-}
-
 func (j JWK) ComputeThumbprint() (string, error) {
-	cJwk := j.ToCJwk()
-	defer j.FreeCJwk(cJwk)
+	cJwk := web5c.NewCJwk(j.ALG, j.KTY, j.CRV, j.D, j.X, j.Y)
+	defer web5c.FreeCJwk(cJwk)
 
-	cThumbprint := C.jwk_compute_thumbprint(cJwk)
-	if cThumbprint == nil {
-		return "", nil
-	}
-	defer C.free_string(cThumbprint)
-
-	thumbprint := C.GoString(cThumbprint)
+	thumbprint := web5c.CJwkComputeThumbprint(cJwk)
 	return thumbprint, nil
+}
+
+func NewJWKFromCJwk(cJwk *web5c.CJwk) *JWK {
+	return &JWK{
+		ALG: cJwk.GetALG(),
+		KTY: cJwk.GetKTY(),
+		CRV: cJwk.GetCRV(),
+		D:   cJwk.GetD(),
+		X:   cJwk.GetX(),
+		Y:   cJwk.GetY(),
+	}
 }
