+++ bound/golang/go.mod
@@ -1,3 +1,3 @@
 module web5
 
-go 1.22.0
+go 1.23.1
