--- bound/golangv2/go.mod
@@ -1,3 +1,0 @@
-module web5
-
-go 1.23.1
